requirements.txt
fastmcp~=2.2.6
python-dotenv~=1.1.0
playwright~=1.52.0
psycopg2-binary
uvicorn~=0.34.2
APScheduler~=3.10.1
